% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{any_char}
\alias{any_char}
\title{Match any character.}
\usage{
any_char(n = 1L, include_newline = FALSE)
}
\arguments{
\item{n}{How many times to match the pattern. See \code{\link{times}}.}

\item{include_newline}{Whether to also match \code{"\\n"}. FALSE by default.}
}
\description{
Match any character.
}
