% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{any_of}
\alias{any_of}
\alias{none_of}
\title{Match any character in a vector.}
\usage{
any_of(characters, n = 1L, negate = FALSE)

none_of(characters, n = 1L, negate = TRUE)
}
\arguments{
\item{characters}{Vector of characters to match.}

\item{n}{How many times to match the pattern. See \link{times}.}

\item{negate}{If TRUE, the pattern will match any character but the ones
in \code{characters}.}
}
\description{
Matches any of the \code{characters}.
}
