% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{capture}
\alias{capture}
\title{Create a capture group.}
\usage{
capture(pattern, name = NULL)
}
\arguments{
\item{name}{Optionally, a string indicating the name of the capture group.}
}
\description{
Encloses the string in parentheses \code{(...)}, and optionally
names the capture group.
}
