% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{maybe}
\alias{maybe}
\title{Maybe match a pattern.}
\usage{
maybe(pattern, nmax = 1L, capture = FALSE)
}
\arguments{
\item{pattern}{String indicating the pattern to match.}

\item{nmax}{Maximum number of times to match the pattern.}

\item{capture}{Logical indicating whether to capture the pattern, or
a string indicating the name of the capture group.}
}
\description{
Alias for \code{times(pattern, n = 0:nmax)}
}
\seealso{
[times()]
}
