% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{word}
\alias{word}
\alias{digit}
\alias{whitespace}
\alias{newline}
\title{Meta Sequences}
\usage{
word(n = 1L, negate = FALSE)

digit(n = 1L, negate = FALSE)

whitespace(n = 1L, negate = FALSE)

newline(n = 1L, negate = FALSE)
}
\arguments{
\item{n}{How many times to match the pattern. See \code{\link{times}}.}

\item{negate}{Match anything but the pattern.}
}
\description{
Meta Sequences
}
\details{
\code{word() == '\w'}. Matches any letter, digit or underscore.
Equivalent to \code{"[a-zA-Z0-9_]"}.

\code{digit() == '\d'}. Matches any digit. Equivalent to \code{"[0-9]"}.

\code{whitespace() == '\s'}. Matches spaces, tabs, and newlines.

\code{newline() == '\n'}. Matches newlines.
}
