% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{option}
\alias{option}
\title{Match one of the patterns.}
\usage{
option(patterns, capture = FALSE)
}
\arguments{
\item{patterns}{String vector of patterns to match.}

\item{capture}{Logical indicating whether to capture the pattern, or
a string indicating the name of the capture group.}
}
\description{
Match any one of the \code{patterns}.
}
